refactor(lowering): extract asm rewrite helpers
