+++ scripts/source-file-size-allowlist.json
@@ -2,6 +2,6 @@
   "hardCap": {
     "src/frontend/parser.ts": 1181,
     "src/lowering/emit.ts": 1180,
-    "src/lowering/functionLowering.ts": 1262
+    "src/lowering/functionLowering.ts": 1014
   }
 }
